requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
Pillow==9.1.1
python-slugify==6.1.2
pytz==2022.1
slugify==0.0.1
sqlparse==0.4.2
text-unidecode==1.3
